+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { Header } from './components';
+import { Header } from './components/header/header';
 import '../src/styles/main.scss';
 import './App.scss';
