+++ src/components/header/header-menu/index.ts
@@ -1,0 +1,1 @@
+export * from './header-menu';
