+++ src/components/header/header-tablet/header-tablet.module.scss
@@ -1,5 +1,4 @@
-@import '../styles/main';
-
+@import '../../../styles/main.scss';
 .header {
   display: none;
   box-shadow: 0 1px 1px 0 #e2e6e9;
