+++ src/components/header/header-tablet/index.ts
@@ -1,2 +1,1 @@
-export * from './header';
 export * from './header-tablet';
